+++ roboticstoolbox/backend/VPython/grid.py
@@ -5,8 +5,8 @@
 
 from vpython import vector, compound, mag, box
 from numpy import sign, ceil, arange
-from roboticstoolbox.backend.VPython.graphics_text import update_grid_numbers
-from roboticstoolbox.backend.VPython.graphics_object2d import Marker2D
+from roboticstoolbox.backend.VPython.text import update_grid_numbers
+from roboticstoolbox.backend.VPython.object2d import Marker2D
 from spatialmath import SE2
