+++ roboticstoolbox/backend/VPython/object2d.py
@@ -4,7 +4,7 @@
 """
 
 from vpython import shapes, radians, extrusion, vector
-from roboticstoolbox.backend.VPython.graphics_stl import import_object_from_numpy_stl
+from roboticstoolbox.backend.VPython.stl import import_object_from_numpy_stl
 
 
 class Object2D:
